models: remove PointerType and OpaqueType
Opaques are a characteristic of the (in)abilities of TypeStore, they
are not part of the type model. e.g. a ContainerType can be opaque.

Pointers are seldom used from Typelib these days *because* they are
close to impossible to manipulate well on the C/C++-to-Ruby boundary.
I bet we won't need them, and we'll just reintroduce them if needed