refactor the specialization handling
The main registration is now done in a properly separated object, which
ensures that we can turn on/off specializations at will, and that
specializations are kept registry-local. One can therefore ensure
isolation of the specializations by e.g. creating a registry,
registering specializations and merging the types into it.