cxx: make tests pass under gccxml
Compared to typelib, gccxml will lack support for vectors of vectors.
GCC-XML reports them as incomplete (that is, without size), which
under Typelib was worked around by setting it to the native size.
modelkit-types wants to explicitely avoid that - to for instance
allow to load for different architectures - so that support is
gone from the gccxml importer.