refactor the castxml/gccxml split
The very ugly code paths that had a 'castxml' boolean flag are gone,
replaced by proper overloading in the CastXMLLoader method.

This reduces code duplication, and makes the explicit selection of
the path to the castxml/gccxml binary a lot more straightforward.

The C++ importer facade also now autodetects whether castxml or
gccxml are present, and selects the relevant one (with a higher
priority for castxml)